Cleaned up code artifacts and bugs.
