+++ Gameplay/entities/enemies/EnemyController.cpp
@@ -1107,13 +1107,12 @@ void Hachiko::Scripting::EnemyController::StartParasiteState()
 	if (_parasite)
 	{
 		_parasite->SetActive(true);
+		_parasite->SetOutlineType(Outline::Type::PRIMARY);
 	}
 
 	_parasite_dropped = true;
 
 	_parasite_dissolving_time_progress = 0;
-
-	_parasite->SetOutlineType(Outline::Type::PRIMARY);
 }
 
 void Hachiko::Scripting::EnemyController::UpdateParasiteState()
@@ -1134,7 +1133,10 @@ Hachiko::Scripting::EnemyState Hachiko::Scripting::EnemyController::TransitionsP
 {
 	if (_parasite_dissolving_time_progress >= _parasite_dissolve_time)
 	{
-		_parasite->SetOutlineType(Outline::Type::NONE);
+		if (_parasite)
+		{
+			_parasite->SetOutlineType(Outline::Type::NONE);
+		}
 
 		return EnemyState::SUPER_DEAD;
 	}
