+++ Source/Engine.vcxproj.filters
@@ -2580,7 +2580,7 @@
     <ClInclude Include="src\events\TimeScaleChangeEventPayload.h" />
     <ClInclude Include="src\core\rendering\SSAOManager.h" />
     <ClInclude Include="src\core\rendering\BlurConfig.h" />
-    <ClInclude Include="src\core\rendering\OutlineType.h" />
+    <ClInclude Include="src\core\rendering\Outline.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Modules">
